+++ HISTORY.rst
@@ -15,6 +15,7 @@ History
 * Removed meta-mixin compatibility code
 * Changed slug size to 255 chars
 * Fixed pagination setting in list views
+* Added API to set default sites if user has permission only for a subset of sites
 
 0.7.0 (2016-03-19)
 ++++++++++++++++++